requirements.txt
dill==0.3.8
pandas==2.2.1
ultralytics==8.1.27
pillow==10.2.0
requests==2.31.0
streamlit==1.32.0
xlrd==2.0.1